Conflicts:
    src/mongo/base/error_codes.yml
    src/mongo/db/SConscript
    src/mongo/db/auth/authorization_manager_impl.cpp
    src/mongo/db/auth/authz_manager_external_state_local.cpp
    src/mongo/db/auth/security_file.cpp
    src/mongo/db/curop.cpp
    src/mongo/db/exec/document_value/value.cpp
    src/mongo/db/mongod_main.cpp
    src/mongo/db/mongod_options.cpp
    src/mongo/db/storage/storage_engine_init.cpp
    src/mongo/db/storage/storage_engine_metadata.cpp
    src/mongo/db/storage/storage_engine_metadata_test.cpp
    src/mongo/db/storage/wiredtiger/wiredtiger_kv_engine.cpp
    src/mongo/db/storage/wiredtiger/wiredtiger_kv_engine.h
    src/mongo/db/storage/wiredtiger/wiredtiger_session_cache.cpp
    src/mongo/dbtests/SConscript
    src/mongo/s/commands/cluster_add_shard_cmd.cpp
    src/mongo/s/mongos_main.cpp
    src/mongo/shell/shell_utils_extended.cpp
    src/mongo/tools/mongobridge_tool/mongobridge_options.cpp
    src/third_party/SConscript

+++ jstests/noPassthroughWithMongod/validate_timeseries_count.js
@@ -7,8 +7,8 @@
  * ]
  */
 
-(function() {
-"use strict";
+import {FeatureFlagUtil} from "jstests/libs/feature_flag_util.js";
+
 let testCount = 0;
 const collNamePrefix = "validate_timeseries_count";
 const bucketNamePrefix = "system.buckets.validate_timeseries_count";
@@ -42,24 +42,28 @@ assert.eq(res.nNonCompliantDocuments, 0);
 assert.eq(res.warnings.length, 0);
 
 // Manually changes the control.count of a version-2 (closed) bucket, expects warnings.
-jsTestLog("Manually changing the 'control.count' of a version-2 bucket.");
-testCount += 1;
-collName = collNamePrefix + testCount;
-bucketName = bucketNamePrefix + testCount;
-db.getCollection(collName).drop();
-assert.commandWorked(db.createCollection(
-    collName, {timeseries: {timeField: "timestamp", metaField: "metadata", granularity: "hours"}}));
-coll = db.getCollection(collName);
-bucket = db.getCollection(bucketName);
-coll.insertMany([...Array(1002).keys()].map(i => ({
-                                                "metadata": {"sensorId": 2, "type": "temperature"},
-                                                "timestamp": ISODate(),
-                                                "temp": i
-                                            })),
-                {ordered: false});
-bucket.updateOne({"meta.sensorId": 2, 'control.version': 2}, {"$set": {"control.count": 10}});
-res = bucket.validate();
-assert(res.valid, tojson(res));
-assert.eq(res.nNonCompliantDocuments, 1);
-assert.eq(res.warnings.length, 1);
-})();
+if (!FeatureFlagUtil.isEnabled(db, "TimeseriesAlwaysUseCompressedBuckets")) {
+    // TODO SERVER-77454: Investigate re-enabling this.
+    jsTestLog("Manually changing the 'control.count' of a version-2 bucket.");
+    testCount += 1;
+    collName = collNamePrefix + testCount;
+    bucketName = bucketNamePrefix + testCount;
+    db.getCollection(collName).drop();
+    assert.commandWorked(db.createCollection(
+        collName,
+        {timeseries: {timeField: "timestamp", metaField: "metadata", granularity: "hours"}}));
+    coll = db.getCollection(collName);
+    bucket = db.getCollection(bucketName);
+    coll.insertMany(
+        [...Array(1002).keys()].map(i => ({
+                                        "metadata": {"sensorId": 2, "type": "temperature"},
+                                        "timestamp": ISODate(),
+                                        "temp": i
+                                    })),
+        {ordered: false});
+    bucket.updateOne({"meta.sensorId": 2, 'control.version': 2}, {"$set": {"control.count": 10}});
+    res = bucket.validate();
+    assert(res.valid, tojson(res));
+    assert.eq(res.nNonCompliantDocuments, 1);
+    assert.eq(res.warnings.length, 1);
+}
